requirements.txt
fastapi
uvicorn
SQLAlchemy
pymysql
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pytest